Refactor to move changes to orttraining
