+++ orttraining/orttraining/training_ops/cuda/tensor/concat.h
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-
+#pragma once
 #include "core/common/common.h"
 #include "core/framework/op_kernel.h"
 #include "core/providers/cuda/cuda_common.h"
@@ -9,9 +9,9 @@
 namespace onnxruntime {
 namespace cuda {
 
-class Concat final : public CudaKernel, public ConcatBase {
+class ConcatTraining final : public CudaKernel, public ConcatBase {
  public:
-  Concat(const OpKernelInfo& info) : CudaKernel(info), ConcatBase(info) {}
+  ConcatTraining(const OpKernelInfo& info) : CudaKernel(info), ConcatBase(info) {}
   Status ComputeInternal(OpKernelContext* context) const override;
 };
